feat: schedule clock-n/out and cancel api is integrated

+++ lib/features/schedule_features/models/active_clock_log_response_model.dart
@@ -15,7 +15,7 @@ class ActiveClockLogResponseModel {
   final String? clockOutAt;
   final String? clockOutLatitude;
   final String? clockOutLongitude;
-  final int? hoursWorked;
+  final double? hoursWorked;
 
   ActiveClockLogResponseModel({
     this.id,
